:ghost: Move Task object into separate file.
